requirements.txt
h5py==2.7.1
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
Keras==2.0.8
torch
typing_extensions
numpy
pandas
pygame
requests
tqdm
